refactor: split demo components and optimize post list
