+++ memory-bank/progress.md
@@ -277,3 +277,4 @@ Successfully completed implementation of conditional Python environment framewor
 - Consider implementing conditional frameworks for Node.js, TypeScript, and web development setups
 - [2025-07-14T06:24:01Z] Task: Added Start Web Dev Server task and Copilot instruction file for launching Next.js dev server from VS Code.
 - [2025-07-14T06:51:41Z] Task: Created web-build.instructions.md, added production build and start tasks, and updated launch.json with Next.js debug configuration to streamline local deployment and debugging.
+- [2025-07-17T03:53:10Z] Split demo components and consolidated actions. Improved post list initialization.
